requirements.txt
apprise==1.9.3
argcomplete==2.0.0
asgiref==3.6.0
astroid==2.14.2
asttokens==2.2.1
attrs==25.3.0
av==14.3.0
Babel==2.10.3
beautifulsoup4==4.11.2
bidict==0.23.1
blinker==1.5
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.8
colorama==0.4.6
colorzero==2.0
contextlib2==0.6.0.post1
contourpy==1.3.2
cryptography==38.0.4
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.3.2
decorator==5.2.1
degirum==0.16.0
degirum_tools @ git+https://github.com/DeGirum/degirum_tools.git@d2d3f70ce2e76515963a824124e9d941b7b56e8a
dill==0.3.6
distlib==0.3.6
distro==1.8.0
distro-info==1.5+deb12u1
dlib==19.24.6
docutils==0.19
executing==2.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
ffmpegcv==0.3.18
filelock==3.9.0
Flask==2.2.2
fonttools==4.57.0
future==0.18.2
gpiozero==2.0.1
h11==0.14.0
hailort==4.20.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
imutils==0.5.4
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isort==5.6.4
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lazy-object-proxy==1.9.0
lgpio==0.2.2.0
libarchive-c==5.2
libevdev==0.5
logilab-common==1.9.8
lxml==4.9.2
Mako==1.2.4.dev0
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mccabe==0.7.0
meson==1.5.1
more-itertools==8.10.0
msgpack==1.1.0
msgpack-numpy==0.4.7.1
mypy==1.0.1
mypy-extensions==0.4.3
netaddr==0.8.0
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
olefile==0.46
opencv-python==4.11.0.86
packaging==25.0
pafy==0.5.5
parso==0.8.3
pexpect==4.8.0
pgzero==1.2
picamera2==0.3.25
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
pillow==10.4.0
platformdirs==2.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycocotools==2.0.8
pycryptodomex==3.11.0
pycups==2.0.1
pygame==2.1.2
Pygments==2.14.0
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.16.2
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
pyrsistent==0.18.1
pyseccomp==0.1.2
pyserial==3.5
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==0.21.0
python-engineio==4.12.0
python-prctl==1.8.1
python-socketio==5.13.0
pytz==2022.7.1
pyudev==0.24.0
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
reportlab==3.6.12
requests==2.32.3
requests-oauthlib==1.3.0
responses==0.18.0
rfc3987==1.3.8
roman==3.3
rpds-py==0.24.0
rpi-lgpio==0.6
RTIMULib==7.2.1
scipy==1.15.2
Send2Trash==1.8.1b0
sense-hat==2.6.0
simple-websocket==1.1.0
simplejpeg==1.8.2
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
soupsieve==2.3.2
spidev==3.5
ssh-import-id==5.10
stack-data==0.6.3
thonny==4.1.4
toml==0.10.2
tomlkit==0.11.7
tqdm==4.67.1
traitlets==5.14.3
twython==3.8.2
types-aiofiles==22.1
types-annoy==1.17
types-appdirs==1.4
types-aws-xray-sdk==2.10
types-babel==2.11
types-backports.ssl-match-hostname==3.7
types-beautifulsoup4==4.11
types-bleach==5.0
types-boto==2.49
types-braintree==4.17
types-cachetools==5.2
types-caldav==0.10
types-certifi==2021.10.8
types-cffi==1.15
types-chardet==5.0
types-chevron==0.14
types-click-spinner==0.1
types-colorama==0.4
types-commonmark==0.9
types-console-menu==0.7
types-contextvars==2.4
types-croniter==1.3
types-cryptography==3.3
types-D3DShot==0.1
types-dateparser==1.1
types-DateTimeRange==1.2
types-decorator==5.1
types-Deprecated==1.2
types-dj-database-url==1.0
types-docopt==0.6
types-docutils==0.19
types-editdistance==0.6
types-emoji==2.1
types-entrypoints==0.4
types-first==2.0
types-flake8-2020==1.7
types-flake8-bugbear==22.10.27
types-flake8-builtins==2.0
types-flake8-docstrings==1.6
types-flake8-plugin-utils==1.3
types-flake8-rst-docstrings==0.2
types-flake8-simplify==0.19
types-flake8-typing-imports==1.14
types-Flask-Cors==3.0
types-Flask-SQLAlchemy==2.5
types-fpdf2==2.5
types-gdb==12.1
types-google-cloud-ndb==1.11
types-hdbcli==2.14
types-html5lib==1.1
types-httplib2==0.21
types-humanfriendly==10.0
types-invoke==1.7
types-JACK-Client==0.5
types-jmespath==1.0
types-jsonschema==4.17
types-keyboard==0.13
types-ldap3==2.9
types-Markdown==3.4
types-mock==4.0
types-mypy-extensions==0.4
types-mysqlclient==2.1
types-oauthlib==3.2
types-openpyxl==3.0
types-opentracing==2.4
types-paho-mqtt==1.6
types-paramiko==2.11
types-parsimonious==0.10
types-passlib==1.7
types-passpy==1.0
types-peewee==3.15
types-pep8-naming==0.13
types-Pillow==9.3
types-playsound==1.3
types-polib==1.1
types-prettytable==3.4
types-protobuf==3.20
types-psutil==5.9
types-psycopg2==2.9
types-pyaudio==0.2
types-PyAutoGUI==0.9
types-pycurl==7.45
types-pyfarmhash==0.3
types-pyflakes==2.5
types-Pygments==2.13
types-pyinstaller==5.6
types-PyMySQL==1.0
types-pynput==1.7
types-pyOpenSSL==22.1
types-pyRFC3339==1.1
types-PyScreeze==0.1
types-pysftp==0.2
types-pytest-lazy-fixture==0.6
types-python-crontab==2.6
types-python-dateutil==2.8
types-python-gflags==3.1
types-python-jose==3.3
types-python-nmap==0.7
types-python-slugify==6.1
types-pytz==2022.6
types-pyvmomi==7.0
types-pywin32==304
types-PyYAML==6.0
types-redis==4.3
types-regex==2022.10.31
types-requests==2.32.0.20250328
types-retry==0.9
types-Send2Trash==1.8
types-setuptools==65.5
types-simplejson==3.17
types-singledispatch==3.7
types-six==1.16
types-slumber==0.7
types-SQLAlchemy==1.4.43
types-stdlib-list==0.8
types-stripe==3.5
types-tabulate==0.9
types-termcolor==1.1
types-toml==0.10
types-toposort==1.7
types-tqdm==4.64
types-tree-sitter==0.20
types-tree-sitter-languages==1.5
types-ttkthemes==3.2
types-typed-ast==1.5
types-tzlocal==4.2
types-ujson==5.5
types-urllib3==1.26
types-vobject==0.9
types-waitress==2.1
types-whatthepatch==1.0
types-xmltodict==0.13
types-xxhash==3.0
types-zxcvbn==4.4
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
v4l2-python3==0.3.5
virtualenv==20.17.1+ds
wadllib==1.3.6
wcwidth==0.2.13
webcolors==1.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==2.2.2
wrapt==1.14.1
wsproto==1.2.0
youtube-dl==2020.12.2
zipp==1.0.0
